+++ projetotcc-back/src/main/java/br/edu/utfpr/pb/projetotcc/repository/CargoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.utfpr.pb.projetotcc.model.Cargo;
 
 public interface CargoRepository
-			extends JpaRepository<Cargo, Integer>{
+			extends JpaRepository<Cargo, Long>{
 
 }
