+++ projetotcc-back/src/main/java/br/edu/utfpr/pb/projetotcc/repository/ProjetoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.utfpr.pb.projetotcc.model.Projeto;
 
-public interface ProjetoRepository extends JpaRepository<Projeto, Integer>{
+public interface ProjetoRepository extends JpaRepository<Projeto, Long>{
 
 }
