+++ projetotcc-back/src/main/java/br/edu/utfpr/pb/projetotcc/service/CargoService.java
@@ -3,6 +3,6 @@ package br.edu.utfpr.pb.projetotcc.service;
 import br.edu.utfpr.pb.projetotcc.model.Cargo;
 
 public interface CargoService
-			extends CrudService<Cargo, Integer>{
+			extends CrudService<Cargo, Long>{
 
 }
