+++ projetotcc-back/src/main/java/br/edu/utfpr/pb/projetotcc/service/ProjetoService.java
@@ -2,6 +2,6 @@ package br.edu.utfpr.pb.projetotcc.service;
 
 import br.edu.utfpr.pb.projetotcc.model.Projeto;
 
-public interface ProjetoService extends CrudService<Projeto, Integer>{
+public interface ProjetoService extends CrudService<Projeto, Long>{
 
 }
